Update Lesson 7 - Calculator.py

+++ Lesson 7 - Calculator.py
@@ -1,6 +1,6 @@
 num1 = input("Enter a number: ")
 num2 = input("Enter a 2nd number: ")
-result = int(num1) + float(num2) #converts the string to a number
+result = int(num1) + int(num2) #converts the string to a number
 
 print(result)
